nukes and the material girl... not me, Madge.
11:09-11:10 pm Eastern 2023-12-08 there was a Definition of a suitcase,
nuclear device in Chamblee, Georgia the Creator fixed it as He does
all of them. The source the KGB, a good friend of the CIA.

I heard it so that means the computers detected it " Look it up" to
quote Madonna who we lost last February To cover up this tragedy.